+++ berrycore/env.sh
@@ -56,7 +56,7 @@ export LC_CTYPE=en_US.UTF-8
 # Welcome Message
 echo ""
 echo "==========================================================="
-echo "     BerryCore v0.81.1 - QNX Extended Userland"
+echo "     BerryCore v0.82 - QNX Extended Userland"
 echo "==========================================================="
 echo ""
 echo "Browse available tools:"